runtime: split app.rs into the src/runtime/app/ module directory
Mechanical split of the 1158-line app.rs into mod.rs (shared aliases,
AppInfo, RustStreamError), service.rs (RustStream builder), run.rs
(run/run_until, drain, signals), scope.rs (BrokerScope and mount tails)
and include.rs (the include family). Internal re-exports keep every
public path byte-identical; struct fields and mount tails became
pub(super) so the sibling submodules can reach them, and doc links that
relied on the old super:: position now use full crate paths.
